fix(#83): 스토리북 catalog prefill 누수/누락 수정
